chore: refactor exporter
